Add PdiCommonUtil
