+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@akashic/akashic-engine",
-  "version": "3.0.0-beta.23",
+  "version": "3.0.0-beta.24",
   "description": "The core library of Akashic Engine",
   "main": "index.js",
   "dependencies": {
